Move colors to utils package
